% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{example_eusilc}
\alias{example_eusilc}
\title{example_eusilc}
\usage{
example_eusilc()
}
\value{
data.frame
}
\description{
Création d'un data.frame jck pour eusilc du package laeken
}
\examples{
# Simplement appeler cette fonction pour créer un jeu de données pour eusilc
df_test_jck <- example_eusilc()
str(df_test_jck)
}
