% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop_my_svymean}
\alias{loop_my_svymean}
\title{loop_my_svymean}
\usage{
loop_my_svymean(design, vars_indics, vars_breaks = NULL, vars_domain = NULL)
}
\arguments{
\item{design}{objet my_svrepdesign}

\item{vars_indics}{vecteur d'indicateurs}

\item{vars_breaks}{vecteur de breakdowns}

\item{vars_domain}{vecteur de domaines}
}
\value{
data.frame, résultats de variances
}
\description{
Boule d'estimations de variance v1 (Approche Statbel)
}
\examples{
eusilc_design <- my_svyrepdesign(data=eusilc,
  data_jck = eusilc_jck,
  var_poids = "rb050",
  var_strate = "db040",
  var_facteur = "SCALE_JCK",
  var_id = "db030",
  var_JCK = "ID_JCK")

my_svymean(eusilc_design,~eqIncome)
my_svymean(eusilc_design,~eqIncome,~rb090)

# Uniquement des variables d'intérêt
loop_my_svymean(eusilc_design,
                c("eqIncome","py120n"))

# Ajout de variables de sous-population
loop_my_svymean(eusilc_design,
                c("eqIncome","py120n"),
                c("pl030","rb090"))

# Ajout de variables de domaine
loop_my_svymean(eusilc_design,
               c("eqIncome","py120n"),
               c("pl030","rb090"),
               "hsize")
}
