% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop_svymean}
\alias{loop_svymean}
\title{loop_svymean}
\usage{
loop_svymean(design, vars_indics, vars_breaks = NULL, vars_domain = NULL)
}
\arguments{
\item{design}{objet svrepdesign}

\item{vars_indics}{vecteur d'indicateurs}

\item{vars_breaks}{vecteur de breakdowns}

\item{vars_domain}{vecteur de domaines}
}
\value{
data.frame, résultats de variances
}
\description{
Boule d'estimations de variances v2 (package survey)
}
\examples{
# Nous devons reproduire les étapes de transformation en données larges pour
# obtenir la matrice des poids répliqués, et obtenir le vecteur des facteurs
# de corrections. Nous avons stocké eusilc_jck sous forme d'une liste de
# ménages*réplications, afin de limiter l'espace disque. Les commandes du
# package survey impose de générer la matrice des poids répliqués pour les
# individus, nous devons donc coupler préalablement à eusilc pour obtenir un
# fichier individus.

repweights_matrix <- eusilc_jck \%>\%
  left_join(eusilc \%>\% select(db030,rb030),
            by="db030",relationship = "many-to-many") \%>\%
  select(db030, rb030, PSU_JCK, rb050) \%>\%
  arrange(PSU_JCK) \%>\%
  pivot_wider(
    id_cols = c(db030,rb030),
    names_from = PSU_JCK,
    values_from = rb050,
    names_prefix = "WEI_REP_",
    values_fill = 0
  ) \%>\%
  select(starts_with("WEI_REP_")) \%>\%
  as.matrix()

scale_vector <- eusilc_jck \%>\% select(PSU_JCK,SCALE_JCK) \%>\%
  arrange(PSU_JCK) \%>\% distinct() \%>\% pull(SCALE_JCK)

eusilc_design <- survey::svrepdesign(
  data = eusilc,
  weights = ~rb050,
  repweights = repweights_matrix,
  scale = 1,
  rscales = scale_vector,
  type="JKn"
)
survey::svymean(~eqIncome,eusilc_design)
survey::svyby(~eqIncome,~rb090,eusilc_design,survey::svymean)

# Uniquement des variables d'intérêt
loop_survey(eusilc_design,
            c("eqIncome","py120n"))

# Ajout de variables de sous-population
loop_survey(eusilc_design,
            c("eqIncome","py120n"),
            c("pl030","rb090"))

# Ajout de variables de domaine
loop_survey(eusilc_design,
            c("eqIncome","py120n"),
            c("pl030","rb090"),
            "hsize")
}
