% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_svy.R
\name{my_svyrepdesign}
\alias{my_svyrepdesign}
\title{my_svyrepdesign}
\usage{
my_svyrepdesign(
  data,
  data_jck,
  var_poids = "RB050",
  var_strate = "STRATUM_JCK",
  var_facteur = "FACTEUR_H",
  var_id = "DB030",
  var_JCK = "ID_JCK"
)
}
\arguments{
\item{data}{data.frame avec les données d'enquête}

\item{data_jck}{data.frame avec les poids répliqués}

\item{var_poids}{variable de poids}

\item{var_strate}{variable de pseudo-strate}

\item{var_facteur}{variable de facteur de correction}

\item{var_id}{variable d'identification de l'individus/ménage}

\item{var_JCK}{variable d'identification de la réplication}
}
\value{
object my_svyrepdesign
}
\description{
Génère un design de sondage spécifique pour un jackknife stratifié
}
\examples{
eusilc_design <- my_svyrepdesign(data=eusilc,
  data_jck = eusilc_jck,
  var_poids = "rb050",
  var_strate = "db040",
  var_facteur = "SCALE_JCK",
  var_id = "db030",
  var_JCK = "ID_JCK")

my_svymean(eusilc_design,~eqIncome)
my_svymean(eusilc_design,~eqIncome,~rb090)
}
