% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceSILC-package.R
\docType{package}
\name{varianceSILC-package}
\alias{varianceSILC}
\alias{varianceSILC-package}
\title{varianceSILC: Estimation de la variance dans SILC avec poids répliqués}
\description{
Fournit des fonctions pour estimer la variance des indicateurs issus de l'enquête SILC à l'aide de poids répliqués construits selon la méthode du jackknife stratifié, en cohérence avec le plan de sondage utilisé par Statbel. Inclut des fonctions alternatives à celles du package 'survey' pour une meilleure prise en compte des pseudo-strates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://statbel.fgov.be}
  \item Report bugs at \url{https://github.com/tdelc/varianceSILC/issues}
}

}
\author{
\strong{Maintainer}: Thomas Delclite \email{thomas.delclite@economie.fgov.be} (Statbel, Institut National de Statistique)

}
\keyword{internal}
